Add the Apache 2.0 license to the templates
Templates will be dual licensed, as I do not care in the slightest if
they are used in proprietary software. The reason for this is that I may
end up using templates to generate documents, like PDF documents. And I
do not want to deal with the AGPL in these cases.